:What I've done is the following:
* Added Menus (File and Edit).    Users Expect Menu/s, Most apps have them, it's compatible, and good for hiding away infreuently needed stuff.

The "File" menu allows you to perform:
File -> Open STL
File -> Open Gcode
File-> Preferences ->Load From...
File-> Preferences ->Save
File-> Preferences ->Save As...

In the "Edit" menu,  I put all the "preferences" and "settings" related stuff, including all sliders, options, etc all into 5 Menu Options under
Edit -> Printer Settings
Edit -> Raft Settings
Edit -> Display Settings
Edit -> Custom Button Setup
Edit -> Calibrate

On the Remaining Tabs  ( which are now much simpler and cleaner ) , I increase the size of the primary buttons ("Load STL" and "Convert to GCODE" , and "Print" ).

In my workflow I find that after I generate GCODE/s, and am about to actually "use" my RepRap,  the next step in the process is to use the Interactive button/s to validate key parts of the printer behavious ( eg home the axis', or head the extruder and extrude a little. ).   As a result of this, and my desire to "see" ( via the Serial comms ) that the printer is "working properly". I've promoted the jogging and manual controls to their own tab.   On that Tab, I've also put the Communications and Logging Tabs that were also hidden away more before.

Finally,  the "Print" tab, is *really* simple, but the Pause and Print buttons are very large, as the pause is something you want to be able to find quickly ( if you have a fault or similar ) .