Make the infill slicer properties functional
Correct the spelling of the InfillDistance glade widget name in
the settings so it gets found and connected correctly. Incidentally
fixes the infinite loop in the GCode conversion. Don't allow the
InfillDistance to be < 0.1, since 0.0 causes the infinite loop.

Add more digits to the float spin buttons on the property page
to show the full accurate value.