Don't run to E0 every time we move
G0 E0 does not mean "no extrude" on steppers in absolute-E mode
It means rewind the feedstock to the 0 position.  This will
cause a failed print most of the time. But the printlines code
is doing this every time it needs to move the head before printing.

Remove the E0 from the "move" code to avoid this dangerous
rewind.

It's possible this has been overlooked because of most of the
code now handling anti-ooze moves, but the Raft code still uses
this printlines method.