gcode: Don't emit E0 in incrementalEcode mode
When moving the head around we decide if we need to extrude anything
based on the 'e' value being the same as before when in incremental
E-code mode.  But sometimes we find that the e value is set to 0
to mean "don't extrude".  It causes great havoc when we send an E0
to the printer to mean "no extruding" but we're in incremental mode.

In almost all cases, we never want to move the extruder to the 0mm
position, so treat E0-attempts as a non-extruding move.

This seems like a nasty hack; somewhere there's a rogue setting
e=0 that's causing the real problem, I suspect.