Commands: rework g-code parser to handle odd-balls
Currently the G-Code parser depends on spaces to delimit
individual parameters.  But spaces aren't required in
the G-Code standard.  Also, some valid code formats
(like "G01") are being ignored.

As an example, the following code does not parse correctly:
  G01 X 25 Y12Z10 ( this is a comment )
  G 0 X 100 Y 20

Update the parser to ignore spaces and handle numeric
parsing of Gxx commands according to the EIA standard.
While we're at it, skip the standard comments style
(parentheses), and run it all as a state machine so we
examine each byte just one time.

Also, clean up a warning line in gcode.cpp.