gitversion.sh again - more fixing for out-of-tree builds
Check the src dir for the existence of .git, not the build dir.
Add an extra include directive to find the generated gitversion.h
in the build dir.
Instead of failing, report git commit and date as 'unknown' in the
about dialog when no other information is available (such as
building from github's zipfile version of the repository)