repsnapper FTBFS due to mismathing headers in linux-libc-dev
The headers in linux-libc-dev are different in powerpc/ppc64el compared
to other archs, they seems to be missing include clauses to the
header under the asm-generic/ directory.
This patch works around that by explicitly including both headers under
asm-generic.
Note that the include of arm/termbits.h had to be removed because it
causes yet another conflict, this time due to the redefinition of types
termios and ktermios, which also only affects powepc/ppc64el.