configure.ac: fix detection of external clipper library
We should use AC_COMPILE_IFELSE to detect if the clipper library is
installed or not. If we use AC_LINK_IFELSE, it requires a link but
there's no main() and there's no -lpolyclipping assigned for it. So
we should change AC_LINK_IFELSE to AC_COMPILE_IFELSE.

Also we add a missing semicolon in body in this commit.

Signed-off-by: Ying-Chun Liu (PaulLiu) <paulliu@debian.org>